renamed and moved to activities folder
